Minor fixes in apkzlib to remove warnings in G3.
Test: not needed

+++ apkzlib/src/main/java/com/android/apkzlib/zip/Eocd.java
@@ -215,7 +215,7 @@ class Eocd {
      * @return the size, in bytes, of the EOCD
      */
     long getEocdSize() {
-        return F_COMMENT_SIZE.endOffset() + comment.length;
+        return (long) F_COMMENT_SIZE.endOffset() + comment.length;
     }
 
     /**
